Create astro_blaster.geo.json
This is the actual geometry of our custom model. It must be placed anywhere within the entity folder. We can generate this in Blockbench relatively simply by optimizing in order to remove redundant faces and generating a compressed template since our model will require a single texture to work with bedrock.