Create item_texture.json
This is our item texture file, in which we define a short name for our texture so that it can be referenced in our render controller. We reference "textures/custom/cosmetics/astro_blaster", as this is where we have placed the texture atlas we compressed in Blockbench. Our shortname is defined as "astro_blaster".