Create astro_blaster.png
This is the compressed texture atlas we created in Blockbench, as multi-texture occupancy tends to be rather difficult to implement in Bedrock, so it is much easier to simply create a single texture like this one. We reference this texture in our item_texture.json file, as well as our player entity definition file, player.entity.json.