Set alpha channel visibility to 1/255 on brewing_stang.png
This is done to prevent the item sprite from displaying in the first person view of the player. It does, however, have the consequence of making item drops invisible, as well as items in item frames. In practice, however, I feel this is a more easily mitigated trade-off as alternatives to item frames, such as armor stands, can be used. Item drops seem to be becoming a more restricted feature anyways, and given the robust trading system, I think this is of minor concern. In any case, this method can be revisited should Bedrock introduct features that allows for either: The control of item sprite generation, or the hiding of the first person generated item sprite (via means other than alpha channel deletion). To be clear, the main takeaway here is that the alpha channel is what is used to generate the item sprite, and we currently have no control over its display in the first person beyond alpha channel deletion, which has other consquences.