Update generic.banking.json
We now have working banking. This method uses the new atan function, and relates angular speed and movement speed to ideal banking angle. Essentially, we are using arctan((ω*v)/g). We also smooth with a tracking variable like the previous implementation of the pitch animation, which can be seen in the entity defination file.