Update blaze.entity.json
This entity file is using a test for a multitexture system. This has been implemented primarily to save on file size, should that become an issue in the future. Essentially, we have expanded the canvas our original texture (autopia_car04 in this case) to include a large blank area. We then created the safe texture, which is the standard saftey placard. This has a blank area at the top, and the content of the image is placed, relatively, where the blank area would be on autopia_car04. When multi-texturing, bedrock acts on a layering system essentially. Materials towards the bottom in the render controller are given display priority (meaning in the event of transparent pixels in both autopia_car04 and safe, in this case, the pixels of safe would be displayed). Note that in order for this to work, we must use the villager_v2_masked material, as only certain materials support texture layering. We have also created a render controller file for our Blaze in the render controllers folder, in which we define the array of textures by the shortnames we have gvien them in this entity defination file (Texture.default and Texture.safe).