Implement actual names in en_US.lang
There really doesn't seem to be any pattern I can discern from the identifiers they chose to use for these... Ended up just pulling another pack off of mcpedl and checking based off of that. It seems to be some weird combonation of the pre and post 1.12 naming schemes, with some random curve balls thrown in.